gave every type of statement its own seperate file.
